requirement.txt
appdirs==1.4.3
certifi==2020.4.5.1
distlib==0.3.0
filelock==3.0.12
numpy==1.20.2
Pillow==8.2.0
pipenv==2020.11.15
six==1.14.0
virtualenv==20.0.18
virtualenv-clone==0.5.4
